Create Zheng_Xu.md
I'm not sure about the joining date because I checked the news.md that Zheng Xu should have joined in our lab on early December. But I checked the random channel and found that he joined in the channel on Dec 25th. I'm not sure about which date should be used...